Add files via upload
HTML, CSS, Bootstrap, and JavaScript were used in the design and development of this front-end web project. The project's main goal is to develop an interactive, fully responsive, and aesthetically pleasing user interface. Mobile-first, completely responsive design for seamless usability on different devices. coding structure that is both modular and maintainable.